+++ lib_xcore_math/src/arch/ref/vect_depth_convert.c
@@ -22,7 +22,7 @@ void vect_s32_to_vect_s16(
     // that the user doesn't have to care about that.
     const right_shift_t b_shr_mod = b_shr - 16;
 
-    for(int k = 0; k < length; k++){
+    for(size_t k = 0; k < length; k++){
         const int32_t B = vlashr32(b[k], b_shr_mod);
         a[k] = vdepth16_32(B);
     }
@@ -36,7 +36,7 @@ void vect_s16_to_vect_s32(
     const int16_t b[],
     const unsigned length)
 {
-    for(int k = 0; k < length; k++){
+    for(size_t k = 0; k < length; k++){
         int16_t B = b[k];
         a[k] = B << 8;
     }
