+++ lib_xcore_math/src/arch/ref/vect_s16_extract.c
@@ -18,7 +18,7 @@ void vect_s16_extract_high_byte(
     const int16_t b[],
     const unsigned len)
 {
-  for(int k = 0; k < len; k++){
+  for(size_t k = 0; k < len; k++){
     a[k] = (b[k] & 0xFF00) >> 8;
   }
 }
@@ -28,7 +28,7 @@ void vect_s16_extract_low_byte(
     const int16_t b[],
     const unsigned len)
 {
-  for(int k = 0; k < len; k++){
+  for(size_t k = 0; k < len; k++){
     a[k] = (b[k] & 0x00FF);
   }
 }
