+++ lib_xcore_math/src/arch/ref/vect_set.c
@@ -15,7 +15,7 @@ void vect_s16_set(
     const int16_t value,
     const unsigned length)
 {
-    for(int i = 0; i < length; i++)
+    for(size_t i = 0; i < length; i++)
         data[i] = value;
 }
 
@@ -26,7 +26,7 @@ void vect_s32_set(
     const int32_t value,
     const unsigned length)
 {
-    for(int i = 0; i < length; i++)
+    for(size_t i = 0; i < length; i++)
         data[i] = value;
 }
 
@@ -38,7 +38,7 @@ void vect_complex_s32_set(
     const int32_t imag_part,
     const unsigned length)
 {
-    for(int i = 0; i < length; i++){
+    for(size_t i = 0; i < length; i++){
         data[i].re = real_part;
         data[i].im = imag_part;
     }
